* commit '816b5df993bf858e15a78e30249b700436fb4f09': (24 commits)
  Explicit lambda return type for clarity.
  Buffer rotation maths fix
  Typo fix
  Clarifying comments
  Additional doxygen documentation.
  Minor comment update.
  Add `RaysQuery.h` to public headers list.
  Bugfix : OpenCL gputil::Buffer fix block on and completion event usage.
  RaysQuery documentation updates
  Changed Query::ranges() to use double precision.
  RaysQuery CPU vs GPU test.
  Query::numberOfRays() fix - was reporting double.
  Re-enable full ray set for test RaysQuery.Gpu
  Corrected pinning mode on Gpu map retry
  OpenCL float cast fix
  Bugfix in gputil::Buffer OpenCL : fix block_on + completion event usage
  Bugfix: incorrect (re)pinning mode for GPU regions buffer.
  Minor adjustment to new voxel upload flags to match existing flags
  Reduce memory copies for RaysQueryGpu
  `RaysQueryGpu` fix
  ...

+++ ohm/Query.h
@@ -83,7 +83,7 @@ public:
   /// to a voxel reported by @c intersectedVoxels().
   ///
   /// @return The array of distances to the line segment for each @c intersectedVoxel().
-  const float *ranges() const;
+  const double *ranges() const;
 
   /// Execute a synchronous query.
   ///
